+++ runtime/core/xnix_runtime_core_cli.c
@@ -2827,6 +2827,8 @@ print_probe(void)
   fputs("\"compatibility_test_type_count\":3,", stdout);
   fputs("\"compatibility_test_result_owner\":\"c\",", stdout);
   fputs("\"compatibility_test_result_type_count\":3,", stdout);
+  fputs("\"execution_readiness_owner\":\"c\",", stdout);
+  fputs("\"execution_readiness_application_count\":1,", stdout);
   fputs("\"desktop_activation_manifest_owner\":\"c\",", stdout);
   fputs("\"desktop_activation_manifest_count\":1,", stdout);
   fputs("\"kde_integration_status_owner\":\"c\",", stdout);
@@ -2920,6 +2922,7 @@ print_write_gate(const char *method_name)
 #include "xnix_runtime_core_cli_compatibility_snapshot_plan.inc"
 #include "xnix_runtime_core_cli_compatibility_test_plan.inc"
 #include "xnix_runtime_core_cli_compatibility_test_result.inc"
+#include "xnix_runtime_core_cli_execution_readiness.inc"
 #include "xnix_runtime_core_cli_compatibility_repair_plan.inc"
 #include "xnix_runtime_core_cli_ai_diagnostic_input.inc"
 #include "xnix_runtime_core_cli_ai_diagnostic_recommendation.inc"