+++ test/test_runtime_contract.rb
@@ -44,6 +44,7 @@ plasmoid = project_root.join("kde/plasmoids/org.xnix.compatibilitycenter/metadat
   GetRepairPlan
   GetTestPlan
   GetTestResult
+  GetExecutionReadiness
   GetAIDiagnosticInput
   GetAIDiagnosticRecommendation
   GetAIRepairApprovalGate